Small refactor to ProductProjection
Previously the product projection search filter worked on products
and we transformed the products to product projection after that. With
this change we first transform all products to product projection based
on if we want to also search staged items and then filter.

This should add the ability to only search on staged or on non staged
products.