Added missing test to cover OrderSetDeliveryCustomFieldAction in the … (#204)
For our project we want to set the status of delivery that is attach to
an order. This field is nested in the shippingInfo property. The tests
are at the moment failing because the case to cover
OrderSetDeliveryCustomFieldAction was missing.
I added this test and also unit test to validate the method.