feat: product search add support for exact with multiple values (#281)
**Context**
Commercetools has added support for passing multiple values in a `exact`
expression in the query of a product search. This has the benefit of
allowing more values to be used then creating separate `or` statements
(see: https://docs.commercetools.com/api/search-query-language#exact).
This adds support for this new way of querying.

**Changes**
- Update commercetools sdk to latest version (only major change is
support dropped for node16)
- Add support for new `values` in `exact` expression.