Add basic support for discount codes in the cart repository (#360)
Currently it is not possible to provide discountcodes on cart creation,
nor use the `addDiscountCode` update action on the cart. As I would like
to use this functionality in our repository I've created this PR.

To avoid putting a lot of work in recreating CommerceTools logic I'm not
validating the to-be-added discountcodes further than purely whether
their object exists in the store. So prices will not be updated nor will
cart-discount rules be checked.

---------

Co-authored-by: Daan Meijers <daan.meijers@yoursurprise.com>